Add support for commandline output (closes #33)
This adds very rudimentary support to print the output of
`wifiscanner` to the commandline. As of now, this is simply a `bin/main.rs`
file that iterates over all networks and prints them in order.

The binary can then be installed with `cargo install` via crates.io
to streamline the user-experience.